fix: config and handle lack of discord config better

+++ Domain/DiscordDomain/DeleteUnusedGroupsInvocable.cs
@@ -16,6 +16,9 @@ public class DeleteUnusedGroupsInvocable(
 
         var scope = serviceProvider.CreateScope();
         var discordService = scope.ServiceProvider.GetRequiredService<DiscordService>();
-        await discordService.DeleteUnusedGroupChannels();
+        if (discordService.IsEnabled)
+        {
+            await discordService.DeleteUnusedGroupChannels();
+        }
     }
 }
